+++ src/wad_stats.h
@@ -47,6 +47,8 @@ void WS_Init(void);
 void WS_Save(void);
 void WS_Cleanup(void);
 
+void WS_ClearMapStats(int i);
+
 void WS_WatchMap(void);
 void WS_UnwatchMap(void);
 void WS_WatchKill(void);